+++ CHANGELOG.md
@@ -10,3 +10,8 @@
 ===
 - Updated data set to include the latest JDK update releases
 - Added "tags" command to show these releases per JDK version
+
+0.4
+===
+- Updated data set to include the latest JDK update releases
+- Add `fakeEndTime` field to mark `endTime` values as fake (equal to `startTime` on purpose)